+++ cmd/skywire-cli/commands/config/gen.go
@@ -13,6 +13,8 @@ import (
 	"github.com/skycoin/skycoin/src/util/logging"
 	"github.com/spf13/cobra"
 
+	"github.com/skycoin/skywire/internal/netutil"
+	"github.com/skycoin/skywire/pkg/app/launcher"
 	"github.com/skycoin/skywire/pkg/skyenv"
 	"github.com/skycoin/skywire/pkg/visor/visorconfig"
 )
@@ -28,10 +30,16 @@ var (
 	replaceHypervisors bool
 	testEnv            bool
 	packageConfig      bool
-	skybianConfig      bool
 	hypervisor         bool
 	hypervisorPKs      string
 	dmsgHTTP           bool
+	publicRPC          bool
+	vpnServerEnable    bool
+	disableAUTH        bool
+	enableAUTH         bool
+	selectedOS         string
+	disableApps        string
+	bestProtocol       bool
 )
 
 func init() {
@@ -40,11 +48,17 @@ func init() {
 	genConfigCmd.Flags().BoolVarP(&replace, "replace", "r", false, "rewrite existing config (retains keys).")
 	genConfigCmd.Flags().BoolVarP(&replaceHypervisors, "use-old-hypervisors", "x", false, "use old hypervisors keys.")
 	genConfigCmd.Flags().BoolVarP(&packageConfig, "package", "p", false, "use defaults for package-based installations in /opt/skywire")
-	genConfigCmd.Flags().BoolVarP(&skybianConfig, "skybian", "s", false, "use defaults paths found in skybian\n writes config to /etc/skywire-config.json")
 	genConfigCmd.Flags().BoolVarP(&testEnv, "testenv", "t", false, "use test deployment service.")
 	genConfigCmd.Flags().BoolVarP(&hypervisor, "is-hypervisor", "i", false, "generate a hypervisor configuration.")
 	genConfigCmd.Flags().StringVar(&hypervisorPKs, "hypervisor-pks", "", "public keys of hypervisors that should be added to this visor")
 	genConfigCmd.Flags().BoolVarP(&dmsgHTTP, "dmsghttp", "d", false, "connect to Skywire Services via dmsg")
+	genConfigCmd.Flags().BoolVar(&publicRPC, "public-rpc", false, "change rpc service to public.")
+	genConfigCmd.Flags().BoolVar(&vpnServerEnable, "vpn-server-enable", false, "enable vpn server in generated config.")
+	genConfigCmd.Flags().BoolVar(&disableAUTH, "disable-auth", false, "disable auth on hypervisor UI.")
+	genConfigCmd.Flags().BoolVar(&enableAUTH, "enable-auth", false, "enable auth on hypervisor UI.")
+	genConfigCmd.Flags().StringVar(&selectedOS, "os", "linux", "generate configuration with paths for 'macos' or 'windows'")
+	genConfigCmd.Flags().StringVar(&disableApps, "disable-apps", "", "set list of apps to disable, separated by ','")
+	genConfigCmd.Flags().BoolVarP(&bestProtocol, "best-protocol", "b", false, "choose best protocol (dmsg / direct) to connect based on location")
 }
 
 var genConfigCmd = &cobra.Command{
@@ -56,23 +70,24 @@ var genConfigCmd = &cobra.Command{
 			logger.WithError(err).Fatal("Invalid output provided.")
 		}
 	},
-	Run: func(_ *cobra.Command, _ []string) {
+	Run: func(cmd *cobra.Command, _ []string) {
 		mLog := logging.NewMasterLogger()
 		mLog.SetLevel(logrus.InfoLevel)
 
-		//Fail on -pst combination
-		if (packageConfig && skybianConfig) || (packageConfig && testEnv) || (skybianConfig && testEnv) {
+		//Fail on -pt combination
+		if packageConfig && testEnv {
 			logger.Fatal("Failed to create config: use of mutually exclusive flags")
 		}
 
 		//set output for package and skybian configs
 		if packageConfig {
-			configName := "skywire-config.json"
-			output = filepath.Join(skyenv.PackageSkywirePath(), configName)
-		}
-
-		if skybianConfig {
-			output = "/etc/skywire-config.json"
+			configName := "skywire-visor.json"
+			if hypervisor {
+				configName = "skywire.json"
+			}
+			if !cmd.Flags().Changed("output") {
+				output = filepath.Join(skyenv.PackageSkywirePath(), configName)
+			}
 		}
 
 		// Read in old config (if any) and obtain old secret key.
@@ -90,8 +105,6 @@ var genConfigCmd = &cobra.Command{
 		//  default paths for different installations
 		if packageConfig {
 			genConf = visorconfig.MakePackageConfig
-		} else if skybianConfig {
-			genConf = visorconfig.MakeDefaultConfig
 		} else if testEnv {
 			genConf = visorconfig.MakeTestConfig
 		} else {
@@ -102,6 +115,35 @@ var genConfigCmd = &cobra.Command{
 		conf, err := genConf(mLog, output, &sk, hypervisor)
 		if err != nil {
 			logger.WithError(err).Fatal("Failed to create config.")
+		}
+
+		// Manipulate Hypervisor PKs
+		if hypervisorPKs != "" {
+			keys := strings.Split(hypervisorPKs, ",")
+			for _, key := range keys {
+				keyParsed, err := coinCipher.PubKeyFromHex(strings.TrimSpace(key))
+				if err != nil {
+					logger.WithError(err).Fatalf("Failed to parse hypervisor private key: %s.", key)
+				}
+				conf.Hypervisors = append(conf.Hypervisors, cipher.PubKey(keyParsed))
+
+				// Compare key value and visor PK, if same, then this visor should be hypervisor
+				if key == conf.PK.Hex() {
+					hypervisor = true
+					conf, err = genConf(mLog, output, &sk, hypervisor)
+					if err != nil {
+						logger.WithError(err).Fatal("Failed to create config.")
+					}
+					conf.Hypervisors = []cipher.PubKey{}
+					break
+				}
+			}
+		}
+
+		if bestProtocol {
+			if netutil.LocalProtocol() {
+				dmsgHTTP = true
+			}
 		}
 
 		// Use dmsg urls for services and add dmsg-servers
@@ -141,14 +183,54 @@ var genConfigCmd = &cobra.Command{
 			}
 		}
 
-		if hypervisorPKs != "" {
-			keys := strings.Split(hypervisorPKs, ",")
-			for _, key := range keys {
-				keyParsed, err := coinCipher.PubKeyFromHex(strings.TrimSpace(key))
-				if err != nil {
-					logger.WithError(err).Fatalf("Failed to parse hypervisor private key: %s.", key)
+		// Change rpc address from local to public
+		if publicRPC {
+			conf.CLIAddr = ":3435"
+		}
+
+		// Set autostart enable for vpn-server
+		if vpnServerEnable {
+			for i, app := range conf.Launcher.Apps {
+				if app.Name == "vpn-server" {
+					conf.Launcher.Apps[i].AutoStart = true
 				}
-				conf.Hypervisors = append(conf.Hypervisors, cipher.PubKey(keyParsed))
+			}
+		}
+
+		// Disable apps that listed on --disable-apps flag
+		if disableApps != "" {
+			apps := strings.Split(disableApps, ",")
+			appsSlice := make(map[string]bool)
+			for _, app := range apps {
+				appsSlice[app] = true
+			}
+			var newConfLauncherApps []launcher.AppConfig
+			for _, app := range conf.Launcher.Apps {
+				if _, ok := appsSlice[app.Name]; !ok {
+					newConfLauncherApps = append(newConfLauncherApps, app)
+				}
+			}
+			conf.Launcher.Apps = newConfLauncherApps
+		}
+
+		// Make false EnableAuth for hypervisor UI by --disable-auth flag
+		if disableAUTH {
+			if hypervisor {
+				conf.Hypervisor.EnableAuth = false
+			}
+		}
+
+		// Make true EnableAuth for hypervisor UI by --enable-auth flag
+		if enableAUTH {
+			if hypervisor {
+				conf.Hypervisor.EnableAuth = true
+			}
+		}
+
+		// Check OS and enable auth for windows or macos
+		if selectedOS == "windows" || selectedOS == "macos" {
+			if hypervisor {
+				conf.Hypervisor.EnableAuth = true
 			}
 		}